+++ controllers/MailController.php
@@ -342,6 +342,24 @@ class MailController extends Controller
     }
 
     /**
+     * Mark a Conversation as unread
+     *
+     * @param int $id
+     */
+    public function actionMarkUnread($id)
+    {
+        $this->forcePostRequest();
+        $this->getMessage($id, true)->markUnread(Yii::$app->user->id);
+
+        $nextReadMessage = $this->getNextReadMessage($id);
+
+        return $this->asJson([
+            'success' => true,
+            'redirect' => $nextReadMessage ? Url::toMessenger($nextReadMessage) : Url::to(['/dashboard'])
+        ]);
+    }
+
+    /**
      * Leave Message / Conversation
      *
      * Leave is only possible when at least two people are in the
@@ -441,18 +459,15 @@ class MailController extends Controller
      *
      * @param int $id
      * @param bool $throw
-     * @return Message
+     * @return Message|null
      * @throws HttpException
      */
-    private function getMessage($id, $throw = false)
+    private function getMessage($id, $throw = false): ?Message
     {
         $message = Message::findOne(['id' => $id]);
 
-        if ($message) {
-            $userMessage = $message->getUserMessage();
-            if ($userMessage != null) {
-                return $message;
-            }
+        if ($message && $message->getUserMessage() !== null) {
+            return $message;
         }
 
         if ($throw) {
@@ -460,5 +475,16 @@ class MailController extends Controller
         }
 
         return null;
+    }
+
+    private function getNextReadMessage($id): ?Message
+    {
+        return Message::find()
+            ->leftJoin('user_message', 'user_message.message_id = message.id')
+            ->where(['user_id' => Yii::$app->user->id])
+            ->andWhere(['!=', 'message_id', $id])
+            ->andWhere(['<=', 'last_viewed', 'updated_at'])
+            ->orderBy(['updated_at' => SORT_DESC])
+            ->one();
     }
 }